fix(device): 性能监控柱状图 Long 没有 .dp 扩展,先转 Int feat(device): WiFi 分析(信号/频段/信道/本机地址,不申请定位权限)

+++ app/src/main/kotlin/com/toolbox/nativetoolbox/ui/tools/PerformanceToolScreen.kt
@@ -221,7 +221,7 @@ fun PerformanceToolScreen(onBack: () -> Unit) {
                                         Modifier
                                             .align(androidx.compose.ui.Alignment.BottomCenter)
                                             .fillMaxWidth()
-                                            .height((60 * value / maxValue).dp.coerceAtLeast(2.dp))
+                                            .height((60 * value / maxValue).toInt().dp.coerceAtLeast(2.dp))
                                             .background(palette.accent.copy(alpha = 0.6f))
                                     )
                                 }
